requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
Django==3.1.8
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-filter==2.4.0
django-summernote==0.8.11.6
djangorestframework==3.12.4
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.3.1
pyparsing==2.4.7
python-barcode==0.13.1
python-dateutil==2.8.2
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.6
